Eliminate circular dependency: merge AsyncResult into result.ts
Move AsyncResult class, async types, _errorMessage helper, and
captureAsyncResult into result.ts and delete asyncResult.ts. This
removes the bidirectional import cycle between result.ts and
asyncResult.ts which could break CommonJS module resolution.

Also fix missing type parameter on fail() in AsyncResult.thenOnSuccess
failure path (fail(r.message) -> fail<TN>(r.message)).
