+++ libraries/ts-utils/src/packlets/base/index.ts
@@ -20,7 +20,6 @@
  * SOFTWARE.
  */
 
-export * from './asyncResult';
 export * from './brand';
 export * from './mapResults';
 export * from './messageAggregator';
